requirements.txt
Django==2.1.2
djangorestframework==3.8.2
flake8==3.5.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
pytz==2018.5
